--- components/home/sections/introDetail.js
@@ -1,38 +1,0 @@
-const introDetail = () => (
-  <section
-    className="page-section pt-0 pb-0 banner-section bg-dark"
-    data-background="/static/images/section-bg-2.jpg"
-  >
-    <div className="container relative">
-      <div className="row">
-        <div className="col-sm-6">
-          <div className="mt-60 mt-lg-60 mb-60 mb-lg-60">
-            <div className="banner-content">
-              <h3 className="banner-heading font-alt">Manifest</h3>
-              <div className="banner-description">
-                But the height of the rim on the 35th street side has always
-                been 11 feet and the height of the of the 36 street side has
-                always been 10 1/2.
-              </div>
-              <br />
-              <div className="banner-description">
-                What I would say is just leave it how it is , because you guys
-                are use to it already . Just protect the court when you see an
-                unfamiliar face , more importantly an unfamiliar 5 , and just
-                win
-              </div>
-            </div>
-            <br />
-            <br />
-            <div className={'align-right'}>Full Speech by Sealy</div>
-          </div>
-        </div>
-        <div className="col-sm-6">
-          <img src="/static/images/promo-1.png" alt="" />
-        </div>
-      </div>
-    </div>
-  </section>
-);
-
-export default introDetail;
