+++ components/layout/index.js
@@ -1,7 +1,7 @@
 import Head from 'next/head';
 import globalStyle from '../../styles/globalStyle.js';
 import Nav from '../nav';
-import FooterSection from '../home/sections/footer';
+import FooterSection from './footer';
 
 export default ({ children, title = 'St.Vartan Hoops' }) => (
   <div>
