+++ components/lifeStyle/data.js
@@ -1,0 +1,34 @@
+const data = [
+    {
+        image: '/static/images/lifestyle/tshirt_white_perfect_shot.jpg',
+        title: 'A Perfect Shot',
+        desc: 'Common ver.',
+    },
+    {
+        image: '/static/images/lifestyle/tshirt_white_perfect_shot_carlos.png',
+        title: 'A Perfect Shot',
+        desc: 'Carlos Ver. "A True Beast!"',
+    },
+    {
+        image: '/static/images/lifestyle/tshirt_white_russ.png',
+        title: 'St.Vartan Tour 2019 White Shirt',
+        desc: 'Russ Ver.',
+    },
+    {
+        image: '/static/images/lifestyle/tshirt_white_sean.png',
+        title: 'St.Vartan Tour 2019 White Shirt',
+        desc: 'Sean Ver. "Space Jam"',
+    },
+    {
+        image: '/static/images/lifestyle/T-Shirt_dark_vallen.jpg',
+        title: 'St.Vartan Tour 2019 Black Shirt',
+        desc: 'Vallen Ver.',
+    },
+    {
+        image: '/static/images/lifestyle/T-Shirt_red_russ.jpg',
+        title: 'St.Vartan Tour 2019 Black Shirt',
+        desc: 'Russ Ver. A Red Variation.',
+    },
+]
+
+export default data;
